remove hard dependency on psycopg2
When used for SQLite, the psycopg2 module is not needed.
Furthermore, added a dependency for argparse (if not already present).
argparse 1.1 is included with Python 2.7; for lower versions, it isn't.
If the import fails, the dependency is added, and argparse will be
installed.

See http://peak.telecommunity.com/DevCenter/setuptools#declaring-extras-optional-features-with-their-own-dependencies.